refactor: HandlerMapping 인터페이스화
