+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handlermapping/HandlerMapping.java
@@ -1,0 +1,10 @@
+package com.interface21.webmvc.servlet.mvc.tobe.handlermapping;
+
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface HandlerMapping {
+
+    void initialize();
+
+    Object getHandler(final HttpServletRequest request);
+}
